Sponsor carousel (#96)
* working toward issue #63

I would love it if someone could go over these to make sure all the
sponsor logos are included before I add a carousel.  "EHS Robotics
Familly Member" still needs to be included because they to not have a
logo.

* reformatting images

changing image size so that the carousel looks uniform

* added sponsor carousel

Added a carousel for the sponsors in the footer but now there is some
white space that needs to be filled.  Also, when either arrow is pressed
it goes to the top of the page, I may add an issue for this.

* sponsor update that was meant to be in the last commit

footer part of the sponsor carousel for more info read previous commit

* fixing the buttons on the sponsor carousel

* fixing the sponsor carousel buttons

fixing the carousel buttons and the GoDaddy typo

* updated and resized logos

small edits

* fixing the shadow in the buttons

* update

commit fix

* merge fix still

still merge fixing

* last commit to fix the merge

* removing unneeded code

* image updates

* updating alt attributes

updating alt attributes to reflect their true purpose

* conflict resolution

* recomended changes

I removed the import that was not being used, changed the button colors
to gray, and change the type in the ospi alt atribute.

* working toward desktop only slider

* test for the desktop sponsor slides

I still need to make it smaller vertically and space the logos out more
but other than that I am pretty happy with it.

* slides for desktop complete

* switched it from css to javascript

Now it should only switch on mobile rather than checking the window
size, it is hard for me to test this though.

* Resolve merge conflicts, remove leftover merge files, and replace JS form factor recognition with Bootstrap .hidden classes

* Change bootstrap classes for bootstrap 3 and modify animation

* Update animation and separate the individual images

* Delete old sponsor images

* Fix play/pause button for sponsor carousel

* Slow down sponsor carousel animation