Fixing footer formating errors
Apparently the <footer> tag was causing some CSS rules to not be applied. As well as using 'row' instead of 'row-fluid' causing the sponsor title to ignore padding rules. The only thing left that needs to be done now is the '#fb-like-footer' rule in our css needs to be removed as it's no longer needed.
Further fixes #95