Resolved #98 in a bit of a better way
Because of how the animation is done in CSS, the images extend across the right side (the parts out of the container hidden by overflow rules). However, this causes the sponsor row to look a little weird, with some padding on the left and no padding on the right. However, the old solution simply took the padding off the entire container, which caused the rest of the content to be off center.